fix(trusty-mpm): satisfy the 500-line file-size cap and doc-comment pointer lint on #4060
Two CI gates were failing on PR #4060 (issue #3981 daemon-held pm_guard
kill-switch flags):

- 500-line file-size cap: `daemon/api.rs`, `daemon/managed_routes/mod.rs`,
  `daemon/managed_routes/lifecycle.rs`, and `session_manager/tests.rs` were
  each over their frozen SLOC budget/cap. Fixed by moving the guard-flags
  router merge to `daemon::serve_http`'s top-level composition point (which
  has SLOC headroom, unlike `api.rs`), moving `ResumeQuery` into
  `guard_flags.rs`, de-duplicating three identical `#3822` project-registration
  call sites in `lifecycle.rs` into one helper, and extracting the by-state
  prune tests out of `session_manager/tests.rs` into a new
  `prune_state_tests.rs` sibling file (mirroring the crate's own established
  `reload_error_tests.rs`/`reactivate_tests.rs` extraction pattern).
- Doc-comment pointer lint: added the four tests the existing `Test:` doc
  comments already cited but that didn't exist yet
  (`guard_flags_response_deserializes`, `guard_flags_route_returns_persisted_values`,
  `guard_flags_route_404s_for_unknown_id`, `resume_route_persists_guard_flags`).

Verified locally: `scripts/check_line_cap.sh` (0 violations),
`scripts/check_test_pointers.sh` (0 dangling pointers), `cargo test -p trusty-mpm`
(0 failed across every binary), `cargo clippy -p trusty-mpm --all-targets`
(clean), `cargo fmt --check` (clean).
