+++ crates/trusty-mpm/src/daemon/managed_routes/guard_flags.rs
@@ -19,13 +19,30 @@
 use std::sync::Arc;
 
 use axum::{
-    Json, extract::Path as AxumPath, extract::State, http::StatusCode, response::IntoResponse,
+    Json, Router,
+    extract::{Path as AxumPath, State},
+    http::StatusCode,
+    response::IntoResponse,
+    routing::get,
 };
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 
 use crate::daemon::state::DaemonState;
+use crate::session_manager::SessionRecord;
 
 use super::summary::parse_id;
+
+/// This route's own tiny sub-router. Merged in at `daemon::serve_http`'s
+/// top-level composition point (NOT inside `api::router` itself, unlike the
+/// `sync_assets`/`provision_status` sub-routers) — `api.rs` is already at its
+/// frozen SLOC budget (`.line-cap-allowlist.tsv`) with zero headroom for even
+/// one more `.merge(...)` chain link, while `daemon::mod` has room to spare.
+pub fn router() -> Router<Arc<DaemonState>> {
+    Router::new().route(
+        "/api/v1/sessions/managed/{id}/guard-flags",
+        get(get_guard_flags),
+    )
+}
 
 /// Response body for `GET /api/v1/sessions/managed/{id}/guard-flags`.
 #[derive(Debug, Serialize, PartialEq, Eq)]
@@ -63,5 +80,109 @@ pub async fn get_guard_flags(
         })
         .into_response(),
         Err(_) => (StatusCode::NOT_FOUND, format!("session {id_str} not found")).into_response(),
+    }
+}
+
+/// Persist operator-captured guard flags after a spawn, updating `record` to
+/// match (so the immediate spawn response already reflects them).
+///
+/// Why: extracted out of `lifecycle::spawn_managed` — that file is already
+/// well over its grandfathered SLOC budget, and this write is a single
+/// self-contained, non-fatal side effect (mirrors the Deliverable-linkage
+/// persist immediately after it in that function). Skips the round trip
+/// entirely when both flags are `false` (the common case — the record
+/// already defaults to `false`/`false` at creation, see
+/// `session_manager::create`), so an ordinary spawn pays no extra store
+/// write.
+/// What: no-op when `disable_hooks`/`pm_unrestricted` are both `false`;
+/// otherwise calls `SessionManager::set_guard_flags`, updating `record` on
+/// success and logging (never failing the spawn) on error.
+/// Test: `guard_flags_persist_on_session` in `session_manager::tests`
+/// (via `SessionManager::set_guard_flags` directly) covers the store
+/// write; this wrapper's skip-when-false and record-sync behavior is
+/// exercised transitively by the HTTP spawn tests in `tests/session_manager_mvp.rs`.
+pub(super) async fn persist_after_spawn(
+    state: &Arc<DaemonState>,
+    record: &mut SessionRecord,
+    disable_hooks: bool,
+    pm_unrestricted: bool,
+) {
+    if !disable_hooks && !pm_unrestricted {
+        return;
+    }
+    match state
+        .session_manager()
+        .await
+        .set_guard_flags(&record.id, disable_hooks, pm_unrestricted)
+        .await
+    {
+        Ok(()) => {
+            record.disable_hooks = disable_hooks;
+            record.pm_unrestricted = pm_unrestricted;
+        }
+        Err(e) => tracing::warn!(
+            id = %record.id,
+            "set_guard_flags failed after spawn: {e}; guard stays fully active"
+        ),
+    }
+}
+
+/// Query params for POST /api/v1/sessions/managed/{id}/resume (issue #3981 Part 2).
+///
+/// Why: `resume` re-spawns the session's PM process, so it is the OTHER
+/// moment (besides spawn) an operator's launching-shell env can legitimately
+/// re-arm or re-confirm `pm_guard`'s kill-switch flags — "resume with the
+/// flag set" is the sanctioned way to disable the guard, per Bob's decision
+/// against a mid-session flip. A query string (not a JSON body) keeps every
+/// existing bodyless `POST .../resume` caller (the GUI, the picker,
+/// programmatic callers) working unchanged — the extractor defaults absent
+/// params to `false`/`false`. That default is DELIBERATELY re-asserted (not
+/// skipped) on every resume: a bypass must be re-declared on the SPECIFIC
+/// `tm sessions resume` call that wants it, never left silently "sticky"
+/// from an earlier resume once an unrelated bodyless resume (GUI, picker)
+/// comes along — that would be the exact sticky-bypass shape this issue
+/// exists to close, just moved from settings.json to the session record.
+/// What: `#[serde(default)]` bools, mirroring `super::DeleteQuery`'s pattern.
+/// Moved here (out of `mod.rs`, review follow-up) alongside
+/// [`persist_after_resume`], the only handler that consumes it.
+/// Test: `resume_route_persists_guard_flags` in `super::tests`.
+#[derive(Debug, Deserialize, Default)]
+pub struct ResumeQuery {
+    #[serde(default)]
+    pub disable_hooks: bool,
+    #[serde(default)]
+    pub pm_unrestricted: bool,
+}
+
+/// Re-assert operator-captured guard flags after a resume, updating `record`
+/// to match.
+///
+/// Why: extracted out of `mod::resume_managed_session` for the same reason
+/// as [`persist_after_spawn`]. Unlike that function, this one is called
+/// UNCONDITIONALLY (never skipped when both flags are `false`) — see
+/// [`ResumeQuery`]'s doc for why a bodyless resume must deliberately
+/// re-arm the guard rather than leaving a prior bypass silently sticky.
+/// What: calls `SessionManager::set_guard_flags`, updating `record` on
+/// success and logging (never failing the resume) on error.
+pub(super) async fn persist_after_resume(
+    state: &Arc<DaemonState>,
+    record: &mut SessionRecord,
+    disable_hooks: bool,
+    pm_unrestricted: bool,
+) {
+    match state
+        .session_manager()
+        .await
+        .set_guard_flags(&record.id, disable_hooks, pm_unrestricted)
+        .await
+    {
+        Ok(()) => {
+            record.disable_hooks = disable_hooks;
+            record.pm_unrestricted = pm_unrestricted;
+        }
+        Err(e) => tracing::warn!(
+            id = %record.id,
+            "set_guard_flags failed after resume: {e}; guard flags unchanged"
+        ),
     }
 }
